apps/worker/requirements.txt
livekit-agents==1.2.11
livekit-plugins-openai==1.2.11
livekit-plugins-elevenlabs==1.2.11
livekit-plugins-silero==1.2.11
httpx==0.28.1
PyJWT>=2.10.0
python-json-logger==3.3.0

